clients: move /maintainer to apps router
