+++ clients/apps/web/src/app/maintainer/[organization]/finance/page.tsx
@@ -1,6 +1,9 @@
+'use client'
+
 import Gatekeeper from '@/components/Dashboard/Gatekeeper/Gatekeeper'
 import Finance from '@/components/Finance/Finance'
 import DashboardLayout from '@/components/Layout/DashboardLayout'
+import { useToast } from '@/components/Toast/use-toast'
 import type { NextLayoutComponentType } from 'next'
 import Head from 'next/head'
 import { useRouter } from 'next/router'
@@ -15,6 +18,8 @@ import { useCurrentOrgAndRepoFromURL } from '../../../../hooks'
 const Page: NextLayoutComponentType = () => {
   const router = useRouter()
   const { org, isLoaded } = useCurrentOrgAndRepoFromURL()
+  const { status } = router.query
+  const { toast } = useToast()
 
   useEffect(() => {
     if (isLoaded && !org) {
@@ -22,6 +27,15 @@ const Page: NextLayoutComponentType = () => {
       return
     }
   }, [isLoaded, org, router])
+
+  useEffect(() => {
+    if (status === 'stripe-connected') {
+      toast({
+        title: 'Stripe setup complete',
+        description: 'Your account is now ready to accept pledges.',
+      })
+    }
+  }, [status, toast])
 
   const pledges = useListPledgesForOrganization(org?.platform, org?.name)
   const rewards = useListRewards(org?.id)
@@ -37,7 +51,7 @@ const Page: NextLayoutComponentType = () => {
           pledges={pledges.data.items}
           rewards={rewards.data.items}
           org={org}
-          tab="contributors"
+          tab="current"
           accounts={accounts.data?.items || []}
         />
       )}
