+++ clients/apps/web/src/app/maintainer/[organization]/finance/rewarded/page.tsx
@@ -1,7 +1,8 @@
+'use client'
+
 import Gatekeeper from '@/components/Dashboard/Gatekeeper/Gatekeeper'
 import Finance from '@/components/Finance/Finance'
 import DashboardLayout from '@/components/Layout/DashboardLayout'
-import { useToast } from '@/components/Toast/use-toast'
 import type { NextLayoutComponentType } from 'next'
 import Head from 'next/head'
 import { useRouter } from 'next/router'
@@ -11,13 +12,11 @@ import {
   useListRewards,
 } from 'polarkit/hooks'
 import { ReactElement, useEffect } from 'react'
-import { useCurrentOrgAndRepoFromURL } from '../../../../hooks'
+import { useCurrentOrgAndRepoFromURL } from '../../../../../hooks'
 
 const Page: NextLayoutComponentType = () => {
   const router = useRouter()
   const { org, isLoaded } = useCurrentOrgAndRepoFromURL()
-  const { status } = router.query
-  const { toast } = useToast()
 
   useEffect(() => {
     if (isLoaded && !org) {
@@ -25,15 +24,6 @@ const Page: NextLayoutComponentType = () => {
       return
     }
   }, [isLoaded, org, router])
-
-  useEffect(() => {
-    if (status === 'stripe-connected') {
-      toast({
-        title: 'Stripe setup complete',
-        description: 'Your account is now ready to accept pledges.',
-      })
-    }
-  }, [status, toast])
 
   const pledges = useListPledgesForOrganization(org?.platform, org?.name)
   const rewards = useListRewards(org?.id)
@@ -49,7 +39,7 @@ const Page: NextLayoutComponentType = () => {
           pledges={pledges.data.items}
           rewards={rewards.data.items}
           org={org}
-          tab="current"
+          tab="rewarded"
           accounts={accounts.data?.items || []}
         />
       )}
