+++ clients/apps/web/src/app/maintainer/layout.tsx
@@ -1,0 +1,9 @@
+import Gatekeeper from '@/components/Dashboard/Gatekeeper/Gatekeeper'
+
+export default function Layout({ children }: { children: React.ReactNode }) {
+  return (
+    <>
+      <Gatekeeper>{children}</Gatekeeper>
+    </>
+  )
+}
