+++ clients/apps/web/src/components/Dashboard/MaintainerRepoSelection.tsx
@@ -1,4 +1,4 @@
-import { useRouter } from 'next/router'
+import { useRouter } from 'next/navigation'
 import { Repository } from 'polarkit/api/client'
 import RepoSelection from '../Organization/RepoSelection'
