+++ clients/apps/web/src/components/Dashboard/filters.ts
@@ -1,4 +1,3 @@
-import { NextRouter } from 'next/router'
 import { IssueListType, IssueSortBy } from 'polarkit/api/client'
 
 export type DashboardFilters = {
@@ -25,47 +24,4 @@ export const DefaultFilters: DashboardFilters = {
   sort: undefined,
   onlyPledged: false,
   onlyBadged: false,
-}
-
-export const navigate = (router: NextRouter, filters: DashboardFilters) => {
-  const params = new URLSearchParams()
-
-  const statuses = []
-  if (filters.statusBacklog) {
-    statuses.push('backlog')
-  }
-  if (filters.statusTriaged) {
-    statuses.push('triaged')
-  }
-  if (filters.statusInProgress) {
-    statuses.push('in_progress')
-  }
-  if (filters.statusPullRequest) {
-    statuses.push('pull_request')
-  }
-  if (filters.statusClosed) {
-    statuses.push('closed')
-  }
-
-  params.set('statuses', statuses.join(','))
-
-  if (filters.q) {
-    params.set('q', filters.q)
-  }
-
-  if (filters.sort) {
-    params.set('sort', filters.sort)
-  }
-
-  if (filters.onlyPledged) {
-    params.set('onlyPledged', '1')
-  }
-
-  if (filters.onlyBadged) {
-    params.set('onlyBadged', '1')
-  }
-
-  const url = new URL(window.location.href)
-  const newPath = `${url.pathname}?${params.toString()}`
-  router.push(url.pathname, newPath)
 }
